cf_update/requirements.txt
python-dotenv
requests
pydantic
cachetools
aiohttp[speedups]
aiocache
aioboto3[s3cse]
git+https://github.com/jazzband/prettytable#egg=prettytable
